requirements.txt
appdirs==1.4.4
black==20.8b1
cffi==1.14.2
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.0
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.14.0
regex==2020.7.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
